requirements.txt
flask
flask-cors
pillow
qrcode
